// src/types/metaData.ts
// types.ts

export interface MetaDataLinks {
  faviconLink32x32: string;
  faviconLink16x16: string;
  faviconLinkAppleTouchIcon: string;
  androidChromeLink192x192: string;
  androidChromeLink512x512: string;
}

export interface MetaDataTitle {
  title: string;
  description: string;
}

export interface MetaDataSocial {
  twitter: string;
  facebook: string;
  linkedin: string;
  instagram: string;
  github: string;
  telegram: string;
  whatsapp: string;
  discord: string;
  slack: string;
}